pyproject.toml
[project]
name = "neural-machine-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flax>=0.12.2",
    "jax>=0.8.2",
    "numpy>=2.4.0",
    "optax>=0.2.6",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sacrebleu>=2.5.1",
    "sentencepiece>=0.2.1",
    "tqdm>=4.67.1",
]
